lighthouse/enrollment-server/requirements.txt
flask
pyyaml
python-dotenv